Refactor tile data types to use Int16 instead of Int32
Switched tile IDs, layer/map dimensions, and coordinates from Int32 to Int16 across Tile, TileInfo, TileLayer, and TileMap. Updated constructors, method signatures, and RenderOrder logic. Removed unused tileset texture retrieval method from TileMap.

+++ src/Ascendance.Domain/Extensions/TileLayerTypeExtensions.cs
@@ -16,7 +16,7 @@ public static class TileLayerTypeExtensions
     /// <returns>The render order value (lower values render first).</returns>
     [System.Runtime.CompilerServices.MethodImpl(
         System.Runtime.CompilerServices.MethodImplOptions.AggressiveInlining)]
-    public static System.Int32 GetRenderOrder(this TileLayerType layerType) => (System.Int32)layerType;
+    public static System.Int16 GetRenderOrder(this TileLayerType layerType) => (System.Int16)layerType;
 
     /// <summary>
     /// Determines whether this layer type should be rendered by default.
